pyproject.toml
[project]
name = "b-route-meter"
version = "1.0.0"
description = "B-Route Smart Meter integration for Home Assistant"
requires-python = ">=3.11"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=custom_components.b_route_meter"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.pylint]
disable = ["import-error", "no-name-in-module"]
ignored-modules = ["homeassistant.core", "homeassistant.helpers.update_coordinator"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-homeassistant-custom-component>=0.13.46",
    "black>=23.0.0",
    "pylint>=2.17.0",
    "coverage>=7.2.0",
]
